+++ services/121-service/src/emails/emails.module.ts
@@ -1,13 +1,11 @@
-import { HttpModule } from '@nestjs/axios';
 import { Module } from '@nestjs/common';
 
-import { AzureGraphTokenService } from '@121-service/src/emails/azure-graph-token.service';
 import { EmailsService } from '@121-service/src/emails/emails.service';
-import { CustomHttpService } from '@121-service/src/shared/services/custom-http.service';
+import { GraphModule } from '@121-service/src/emails/graph/graph.module';
 
 @Module({
-  imports: [HttpModule],
-  providers: [EmailsService, AzureGraphTokenService, CustomHttpService],
+  imports: [GraphModule],
+  providers: [EmailsService],
   controllers: [],
   exports: [EmailsService],
 })
