+++ services/121-service/src/emails/graph/interfaces/graph-file-attachment.interface.ts
@@ -1,0 +1,5 @@
+export interface GraphFileAttachment {
+  '@odata.type': '#microsoft.graph.fileAttachment';
+  name: string;
+  contentBytes: string;
+}
