+++ services/121-service/src/emails/graph/interfaces/graph-recipient.interface.ts
@@ -1,0 +1,3 @@
+export interface GraphRecipient {
+  emailAddress: { address: string };
+}
