+++ services/121-service/src/emails/graph/interfaces/graph-send-mail-request.interface.ts
@@ -1,0 +1,6 @@
+import { GraphMessage } from '@121-service/src/emails/graph/interfaces/graph-message.interface';
+
+export interface GraphSendMailRequest {
+  message: GraphMessage;
+  saveToSentItems: boolean;
+}
